requirements.txt
gunicorn==20.0.4
python-dotenv==0.12.0
